backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
h5py
numpy
scikit-learn
pdfplumber
pytesseract
Pillow
python-docx
